Use even bigger numbers

+++ grakn-graql/src/main/java/ai/grakn/graql/internal/gremlin/fragment/Fragment.java
@@ -96,5 +96,5 @@ public interface Fragment {
         return false;
     }
 
-    long fragmentCost(long previousCost);
+    double fragmentCost(double previousCost);
 }
