+++ grakn-graql/src/main/java/ai/grakn/graql/internal/gremlin/fragment/IsAbstractFragment.java
@@ -39,7 +39,7 @@ class IsAbstractFragment extends AbstractFragment {
     }
 
     @Override
-    public long fragmentCost(long previousCost) {
+    public double fragmentCost(double previousCost) {
         return previousCost;
     }
 }
